Ver2

+++ src/main/java/utilities/ExcelUtil/ExcelWriter.java
@@ -1,4 +1,4 @@
-package utilities;
+package utilities.ExcelUtil;
 
 import Environement.Configuration;
 import org.apache.commons.io.FileUtils;
@@ -19,14 +19,15 @@ public class ExcelWriter {
     private String excelFilePath;
 
     public ExcelWriter(String excelFilePath) {
-
         String[] element = excelFilePath.split("/");
         String fileName = element[element.length-1];
         //Copy file to report folder
         File sourceFile = new File(excelFilePath);
         File destFile = new File(Configuration.getInstance().getReportPath() + fileName);
         try{
-            FileUtils.forceDelete(destFile);
+            if(destFile.exists()){
+                FileUtils.forceDelete(destFile);
+            }
             FileUtils.copyFile(sourceFile, destFile);
             this.excelFilePath = Configuration.getInstance().getReportPath() + fileName;
         } catch (Exception e){
@@ -36,6 +37,13 @@ public class ExcelWriter {
 
     }
 
+    /**
+     * Put all data into a HashMap
+     * @param testIndex
+     * @param result
+     * @param executeTime
+     * @return Hashmap data
+     */
     private HashMap<Integer, List<Object>> putDataIntoMap(int testIndex, List<Object> result, double executeTime){
         HashMap<Integer, List<Object>> map = new HashMap<>();
         result.add(executeTime);
@@ -43,27 +51,34 @@ public class ExcelWriter {
         return map;
     }
 
+    /**
+     * Write the result into the report file
+     * @param testIndex
+     * @param result
+     * @param executeTime
+     */
     public void writeReport(int testIndex, List result, double executeTime) {
         try {
+            //Prepare the input data
             Map<Integer, List<Object>> data = putDataIntoMap(testIndex, result, executeTime);
-
-
-
 
             FileInputStream fIPS= new FileInputStream(excelFilePath); //Read the spreadsheet that needs to be updated
             Workbook wb;
             Sheet worksheet;
             wb = getWorkbook(fIPS, excelFilePath); //If there is already data in a workbook
             worksheet = wb.getSheet("Data");
-            //Access the worksheet, so that we can update / modify it
+
             Row row;
             Cell cell;
             for(Map.Entry<Integer, List<Object>> dataValue : data.entrySet()) {
                 row = worksheet.getRow(dataValue.getKey());
+                //Actual result
                 row.getCell(3).setCellType(CellType.STRING);
                 row.getCell(3).setCellValue(String.valueOf(dataValue.getValue().get(1)));
+                //Status
                 row.getCell(5).setCellType(CellType.STRING);
                 row.getCell(5).setCellValue(String.valueOf(dataValue.getValue().get(0)));
+                //Execute time
                 row.getCell(6).setCellType(CellType.NUMERIC);
                 row.getCell(6).setCellValue(Double.valueOf(String.valueOf(dataValue.getValue().get(2))));
             }
@@ -77,9 +92,15 @@ public class ExcelWriter {
         }
     }
 
+    /**
+     * Get the workbook
+     * @param inputStream
+     * @param excelFilePath
+     * @return the correct workbook
+     * @throws IOException
+     */
     private Workbook getWorkbook(FileInputStream inputStream, String excelFilePath) throws IOException {
         Workbook workbook;
-
         if (excelFilePath.endsWith("xlsx")) {
             workbook = new XSSFWorkbook(inputStream);
         } else if (excelFilePath.endsWith("xls")) {
